Fixed Patricia
Reduced input size.
Switched to stdint types.
Use alloca instead of malloc.
Upgrade to Python3.

+++ patricia/patricia.h
@@ -25,6 +25,7 @@
 #ifndef _PATRICIA_H_
 #define _PATRICIA_H_
 
+#include <stdint.h>
 
 /*
  * Patricia tree mask.
@@ -32,7 +33,7 @@
  * structure is where the mask and data are kept.
  */
 struct ptree_mask {
-	unsigned long pm_mask;
+	uint32_t pm_mask;
 	void *pm_data;
 };
 
@@ -41,7 +42,7 @@ struct ptree_mask {
  * Patricia tree node.
  */
 struct ptree {
-	unsigned long p_key;		/* Node key		*/
+	uint32_t p_key;		/* Node key		*/
 	struct ptree_mask *p_m;		/* Node masks		*/
 	unsigned char p_mlen;		/* Number of masks	*/
 	char p_b;			/* Bit to check		*/
@@ -52,6 +53,6 @@ struct ptree {
 
 extern struct ptree *pat_insert(struct ptree *n, struct ptree *head);
 extern int           pat_remove(struct ptree *n, struct ptree *head);
-extern struct ptree *pat_search(unsigned long key, struct ptree *head);
+extern struct ptree *pat_search(uint32_t key, struct ptree *head);
 
 #endif /* _PATRICIA_H_ */
